Current version of universality_of_hardness.py
It finds the 10 samples from each set type that are the hardest according to the networks. After that it shows the threshold-accuracy graph. In next iteration I plan to modify the code responsible for thresholds. Currently, it uses requires the samples to overlap across all networks, which is suboptimal (doesn't show the whole picture and slightly downplays/overplays the results).