Created measuring_imbalance_of_original_datasets.py. This program will be responsible for the first experiment, in which we show that even on balanced datasets not all classes have the same average accuracies. We next show how this distribution of errors across classes correlates to different metrics of hardness in order to show that hardness-based within-class data imbalance is responsible for those imbalances in balanced datasets. In the future, we should add learning-based metrics for hardness, as they are more disconnected from accuracy making the argument even stronger.
We also modify save_data from utils so that it always saves on CPU, add a function for loading dataset that allows for working on all data (not just training set), and a function for measuring class-level accuracies.

Finally, we add more hardness metrics to compute_confidences.py and appropriately modify the rest of that program.